--- api/models/config_model.go
@@ -1,6 +1,0 @@
-package models
-
-// Config -
-type Config struct {
-	LedgerStorage *LedgerStorage `json:"storage"`
-}
