--- api/models/infos_model.go
@@ -1,8 +1,0 @@
-package models
-
-// Infos -
-type Infos struct {
-	Server  string      `json:"server"`
-	Version interface{} `json:"version"`
-	Config  *Config     `json:"config"`
-}
