--- api/models/ledger_storage_model.go
@@ -1,7 +1,0 @@
-package models
-
-// LedgerStorage -
-type LedgerStorage struct {
-	Driver  interface{} `json:"driver"`
-	Ledgers interface{} `json:"ledgers"`
-}
